+++ source/pijamas/exception.d
@@ -5,41 +5,63 @@
  */
 module pijamas.exception;
 
-//version(Have_unit_threaded)
-static if (__traits(compiles, { import unit_threaded.should : UnitTestException; }))
-{
-  import unit_threaded.should : UnitTestException;
+version (MirNoGCException) {
+  import mir.exception : MirException;
 
-  alias BaseException = UnitTestException;
-
-} else static if (__traits(compiles, { import dunit.error : DUnitAssertError; }))
-{
-  import dunit.error : DUnitAssertError;
-
-  alias BaseException = DUnitAssertError;
+  alias BaseException = MirException;
 
 } else {
+  static if (__traits(compiles, { import unit_threaded.should : UnitTestException; })) {
+    import unit_threaded.should : UnitTestException;
 
-  alias BaseException = Exception;
+    alias BaseException = UnitTestException;
+
+  } else static if (__traits(compiles, { import dunit.error : DUnitAssertError; })) {
+    import dunit.error : DUnitAssertError;
+
+    alias BaseException = DUnitAssertError;
+
+  } else {
+
+    alias BaseException = Exception;
+  }
 }
 
 /**
  * An exception thrown when a test fails
  */
-public class AssertException : BaseException
-{
+public class AssertException : BaseException {
 
-  /**
-   * Constructor
-   *
-   * Params:
-   *  msg = The failure message
-   *  file = The filename where the test failed
-   *  line = The line where the test failed
-   */
-  this(string msg, string file = __FILE__, size_t line = __LINE__) @safe pure
-  {
-    super(msg, file, line);
+  version (MirNoGCException) {
+
+    /**
+    * Constructor
+    *
+    * Params:
+    *  msg = The failure message
+    *  file = The filename where the test failed
+    *  line = The line where the test failed
+    */
+    @nogc @safe pure nothrow this(scope const(char)[] msg, string file = __FILE__, size_t line = __LINE__) {
+      super(msg, file, line);
+    }
+
+    /// ditto
+    @nogc @safe pure nothrow this(string msg, string file = __FILE__, size_t line = __LINE__) {
+      super(msg, file, line);
+    }
+
+  } else {
+    /**
+    * Constructor
+    *
+    * Params:
+    *  msg = The failure message
+    *  file = The filename where the test failed
+    *  line = The line where the test failed
+    */
+    this(string msg, string file = __FILE__, size_t line = __LINE__) @safe pure {
+      super(msg, file, line);
+    }
   }
 }
-